+++ internal/ethapi/api_test.go
@@ -633,7 +633,7 @@ func (b testBackend) SubscribeNewTxsEvent(events chan<- core.NewTxsEvent) event.
 }
 func (b testBackend) ChainConfig() *params.ChainConfig { return b.chain.Config() }
 func (b testBackend) Engine() consensus.Engine         { return b.chain.Engine() }
-func (b testBackend) CurrentTurnTerm() (uint64, error) { return 1, nil }
+func (b testBackend) CurrentTurnTerm() (uint8, error)  { return 1, nil }
 func (b testBackend) GetLogs(ctx context.Context, blockHash common.Hash, number uint64) ([][]*types.Log, error) {
 	panic("implement me")
 }
